fix: [STO-12076]: move sast rem diff extractor to extractors.ts
Architecture guardrail caps inline responseExtractor count in sto.ts.
